+++ api/accounts/urls.py
@@ -1,0 +1,9 @@
+from django.conf.urls import url
+from rest_framework_jwt.views import obtain_jwt_token, refresh_jwt_token
+from .views import register
+
+urlpatterns = [
+    url(r'^login/', obtain_jwt_token),
+    url(r'^token-refresh/', refresh_jwt_token),
+	url(r'^register/', register, name="register")
+]
